+++ lib/src/src.dart
@@ -2,4 +2,5 @@ export 'anchored_animation.dart';
 export 'anchored_controller.dart';
 export 'anchored_drag.dart';
 export 'anchored_flutter_sheet.dart';
+export 'anchored_modal_manager.dart';
 export 'layout/layout.dart';